Split tests out into separate files.
