+++ .travis.yml
@@ -14,7 +14,7 @@ before_install:
 
 script:
   - cp config/example.settings.yml config/settings.dev.yml
-  - php vendor/bin/phpunit --group UnitTest --coverage-clover=coverage.xml
+  - php vendor/bin/phpunit --group UnitTest 
   - php vendor/bin/phpcs --standard=PSR2 src test
 
 notifications:
