+++ test/TransactionService/Controller/RollbackTransactionTest.php
@@ -1,0 +1,44 @@
+<?php
+
+namespace Islandora\Crayfish\Test\TransactionService\Controller;
+
+use Islandora\Chullo\FedoraApi;
+use Islandora\Chullo\TriplestoreClient;
+use Symfony\Component\HttpFoundation\Response;
+use Islandora\Crayfish\Test\CrayfishWebTestCase;
+
+class RollbackTransactionTest extends CrayfishWebTestCase
+{
+    public function setUp()
+    {
+        parent::setUp();
+    }
+    
+    public function createApplication()
+    {
+        return parent::createApplication();
+    }
+
+    /**
+     * @group UnitTest
+     * @covers \Islandora\Crayfish\TransactionService\Controller\TransactionController::rollback
+     */
+    public function testRollbackTransactionOk()
+    {
+        $txID = "tx:f218d271-98ee-4a90-a06a-03420a96d5af";
+        $location = "http://localhost:8080/fcrepo/rest/$txID";
+        $headers = array(
+            'Server' => CrayfishWebTestCase::$serverHeader,
+            'Location' => $location,
+            'Date' => CrayfishWebTestCase::$today,
+        );
+
+        $responseOK = Response::create('', 204, $headers);
+        
+        $this->api->method('rollbackTransaction')->willReturn($responseOK);
+        
+        $client = $this->createClient();
+        $crawler = $client->request('POST', "/islandora/transaction/${txID}/rollback");
+        $this->assertEquals($client->getResponse()->getStatusCode(), 204, "Did not commit transaction.");
+    }
+}
