+++ test/bootstrap.php
@@ -1,2 +1,3 @@
 <?php
-    include __DIR__ . '/../vendor/autoload.php'; // composer autoload
+    include_once __DIR__ . '/../vendor/autoload.php'; // composer autoload
+    include_once __DIR__ . '/CrayfishWebTestCase.php'; // Common application functions
